% Generated by roxygen2 (4.0.2): do not edit by hand
\name{age.sex.zip}
\alias{age.sex.zip}
\title{Simulate CRC data with age, sex, and zip code}
\usage{
age.sex.zip(n)
}
\arguments{
\item{n}{The number of population units.}
}
\value{
\item{pop}{ A data.frame containing the covariates and capture
histories associated with each observed unit. }
}
\description{
Generate a fake data set with covariates age, sex, and zip code.
}
\details{
This simulation is vaguely inspired by the French dataset on multiple
sclerosis prevalence (El Adssi et. al., 2012).  It is used to create a
three-list capture-recapture system with unit-level independence between
lists.  Capture probabilities are constant (for simplicity!) across lists,
but not across population units.  The heterogeneity is a function of the
covariates: age, sex, and zip code.  We code zip here as `a', `b', `c', `d'.
}
\author{
Zach Kurtz
}

