% Generated by roxygen2 (4.0.2): do not edit by hand
\name{french.1}
\alias{french.1}
\title{A fake dataset, french.1}
\usage{
french.1(pop, n, k)
}
\arguments{
\item{pop}{A data frame containing a template for the population.}

\item{n}{The size of the target population.}

\item{k}{The number of lists.}
}
\value{
A data frame, which gets passed back up to the main function
\code{poptop} for a bit of cleaning.
}
\description{
Generate a fake dataset, for simulation purposes, that is somewhat similar
to the data on multiple sclerosis in the Lorraine Region of France as in El
Adssi et. al. (2012).
}
\details{
This function is the guts for a capture-recapture experiment that is
specified by the \code{heters} argument in the function \code{poptop}; with
a minor revision ot the function \code{captures}, you could create your own
guts function for a very different capture-recapture experiment that sits
within the same coding framework.
}
\author{
Zach Kurtz
}
\references{
El Adssi H, Debouverie M and Guillemin F (2012). "Estimating the
prevalence and incidence of multiple sclerosis in the Lorraine region,
France, by the capture-Recapture method." \emph{Multiple Sclerosis Journal},
\bold{18}(9), pp. 1244-1250.
}

