% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resample.captures}
\alias{resample.captures}
\title{Tool for bootstrapping}
\usage{
resample.captures(boot.list)
}
\arguments{
\item{boot.list}{See \code{\link{lllcrc.boots}} or \code{\link{vgam.crc.boots}}.}
}
\description{
Used by \code{vgam.crc.boots} and \code{lllcrc.boots}.  See these functions
for details.
}
\author{
Zach Kurtz
}
\references{
Zwane E and Heijden Pvd (2003). "Implementing the parametric bootstrap
in capture-recapture models with continuous covariates." \emph{Statistics &
Probability Letters}, \bold{65}, pp. 121-125.
}

