% Generated by roxygen2 (4.0.2): do not edit by hand
\name{string.to.array}
\alias{string.to.array}
\title{Put LLM vector into a LLM design matrix}
\usage{
string.to.array(densi, rasch = FALSE)
}
\arguments{
\item{densi}{The capture pattern counts}

\item{rasch}{Logical: if TRUE, the square of the number of captures is
included in the design matrix as the column \code{ck2}}
}
\value{
A design matrix
}
\description{
Given the output of \code{pop.to.counts}, i.e., a 1-row dataframe with as
many columns as there are observable capture patterns, create a design
matrix for log-linear modeling
}
\author{
Zach Kurtz
}

