% Generated by roxygen2 (4.0.2): do not edit by hand
\name{patterns.possible}
\alias{patterns.possible}
\title{Generate all observable capture patterns}
\usage{
patterns.possible(k)
}
\arguments{
\item{k}{The number of lists}
}
\description{
Returns a vector of binary strings representing all possible capture
patterns excluding complete noncapture
}
\author{
Zach Kurtz
}

